Build container image with github actions (#1268)
Just the bare minimum to build the same container image that we got from
cloud build via a github action and push it to
`ghcr.io/embarkstudios/quilkin`. I've bypassed `build/Makefile` entirely
and only picked out the relevant steps, but leaving all the old stuff as
is for now. I created a new `Dockerfile.ghaction` that runs some of the
old make targets in a dockerfile multi stage build instead as that works
better with `docker/build-push-action`.

The `build-env` image is not saved in a remote registry but should be
cached between runs in the gha cache.

The action should now run on
* any push to main
* any new tag (without `/` in the name)
* triggering the workflow manually (can't test this until the workflow
is on main)

One image pushed so far:
https://github.com/EmbarkStudios/quilkin/pkgs/container/quilkin, same
size as the last cloud build image and still has the commit sha embedded
in the binary.