Remove unused version field of EndpointSet, hash is used for version() (#1272)
This field was confusing and probably just a leftover. This is good
though because this means that an EndpointSet is deterministically
versioned and if e.g. two agents are connected to the relay from the
same cluster they will report the same version. I will need to confirm
things but this should mean that we don't need to track "who is
responsible" for a specific locality and lock it to only one agent at a
time.