Add persistent UDP stream (#1271)
Follow up on #1266 which never got any reviews, so this PR is a bit
bulky.

In addition to the things in #1266, this adds a quinn-based UDP server
and client implementation that is meant for agents (clients) to use to
connect to relays (servers) so that the relay can track the connection
status of each agent. This essentially replaces the usage of corrosion's
client implementation for mutations, as that would have needed a
parallel channel so that the relays could determine the up status of
each agent, so this just combines the connection with the mutations
themselves.

The DB schema has changes, as the `servers` table has 2 new columns,
`contributors` and `cont_time`. `contributors` is a JSONB object where
each key is the IP address of an agent that has contributed to it. When
the connection to an agent is lost, it is removed as a contributor from
all the servers. We can then run a garbage collection task that can
remove servers with no contributors, if they have not been updated in
whatever amount of time we want to.

This adds some basic tests for the various DB operations and the quinn
stream.