fix: Ensure Datacenter xDS resource names and versions are consistent (#1276)
Because we were using different values for `XdsResource.name` and the
first parameter to the `client_state.version_matches()` call we would
always think that the version does not match and push an update.

Also make sure we check the version when propagating self.

This exposed a bug where if the relay restarts, a proxy would get stuck
in unhealthy because the xDS client would only move back to healthy if
it received an update, but since the unnecessary updates were removed we
would not get one if nothing had changed.