Remove `LocalVersions.clear()` (#1278)
This changes behavior so that we won't remove all resources that were
added by a given agent from the relay Config when that agent disconnects
gracefully. This could be disrupting in the usual case where the agent
disconnects simply due to a rollout or rescheduling event.

The LocalVersion struct exits scope when the connection is closed so we
don't need to clear the local state it has either.

Since the Cluster (EndpointSet) resource is versioned deterministically
by hashing the endpoints, and updated as a batch, we don't have to worry
about old endpoints lingering if we miss an update when one agent shuts
down before another picks up.

In the case of the Datacenter resource this is handled separately
through the `client_disconnected()` method on the `Configuration` trait,
which in the case of a relay will remove the disconnecting agent from
its DatacenterMap. We also need to make sure we remove the client from
ClusterMap.localities so that the next agent can take over, so I've
added the `remove_contributor()` method that is called in
`client_disconnected()`.